fix(api): isolate lifecycle storage teardown
