+++ turbo/apps/api/src/signals/services/agent-instructions-storage.service.ts
@@ -3,12 +3,7 @@ import {
   getInstructionsFilename,
   SUPPORTED_FRAMEWORKS,
 } from "@okouai/core/frameworks";
-import {
-  getInstructionsStorageName,
-  VOLUME_ORG_USER_ID,
-} from "@okouai/core/storage-names";
-import { storages } from "@okouai/db/schema/storage";
-import { and, eq } from "drizzle-orm";
+import { getInstructionsStorageName } from "@okouai/core/storage-names";
 
 import type { Tx } from "../../lib/db-types";
 import { env } from "../../lib/env";
@@ -20,6 +15,7 @@ import {
   prepareVolumeServerSideWithDb$,
 } from "./storage-volume-publication.service";
 import { uploadVolumeServerSide$ } from "./storage-volume-upload.service";
+import { removeAgentInstructionsStorageInTransaction } from "./agent-instructions-storage-transaction.service";
 
 interface WriteAgentInstructionsStorageArgs {
   readonly orgId: string;
@@ -102,31 +98,6 @@ export const writeAgentInstructionsStorageInTransaction$ = command(
     signal.throwIfAborted();
   },
 );
-
-export async function removeAgentInstructionsStorageInTransaction(
-  tx: Tx,
-  args: { readonly orgId: string; readonly agentName: string },
-): Promise<string | null> {
-  const storageName = getInstructionsStorageName(args.agentName);
-  const [storage] = await tx
-    .select({ id: storages.id, s3Prefix: storages.s3Prefix })
-    .from(storages)
-    .where(
-      and(
-        eq(storages.orgId, args.orgId),
-        eq(storages.userId, VOLUME_ORG_USER_ID),
-        eq(storages.name, storageName),
-      ),
-    )
-    .limit(1);
-
-  if (!storage) {
-    return null;
-  }
-
-  await tx.delete(storages).where(eq(storages.id, storage.id));
-  return storage.s3Prefix;
-}
 
 export const deleteAgentInstructionsStorage$ = command(
   async (
